heart_stroke_risk/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
numpy==1.22.4
pandas==1.4.2
scikit_learn==1.1.1
xgboost==1.6.1
